[ICU-10379] fixes for IBM i - passes all tests! X-Commit-URL: https://ssl.icu-project.org/trac/changeset/34348 Former-commit-id: d031d30541c6a0414ee8a8be25e0a841e8d290af

+++ icu4c/source/test/intltest/apicoll.cpp
@@ -2290,7 +2290,7 @@ void CollationAPITest::TestIterNumeric() {
     uiter_setUTF8(&iter40, "\x34\x30", 2);
     uiter_setUTF8(&iter72, "\x37\x32", 2);
     UCollationResult result = coll.compare(iter40, iter72, errorCode);
-    assertEquals("40<72", UCOL_LESS, result);
+    assertEquals("40<72", (int32_t)UCOL_LESS, (int32_t)result);
 }
 
  void CollationAPITest::dump(UnicodeString msg, RuleBasedCollator* c, UErrorCode& status) {
